Refactor crud queryable interfaces out to separate files
